Replace libsass with dart sass
Fixes #275

+++ db/schema.rb
@@ -34,7 +34,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_06_20_213203) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.bigint "blob_id", null: false
+    t.integer "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
